Configured search

+++ xml-service-pretraga/src/main/java/com/example/demo/dto/PretragaDTO.java
@@ -28,7 +28,7 @@ public class PretragaDTO {
 
     private double kilometrazaDTO;
 
-    private int cijenaDTO;
+    private double cijenaDTO;
 
     private double ocjenaDTO;
